[Upd] moved window to UI lib
