+++ src/UI/CMakeLists.txt
@@ -3,6 +3,7 @@ add_library(
     src/button.cpp
     src/container.cpp
     src/text_input.cpp
+    src/window.cpp
 )
 
 target_include_directories(UserInterface PUBLIC
